fix layout update restart flow
